fix: Amend pipeline errors
- amend worflow ith the correct PAT token for module access
- move WalletSdk provider into its own module
- update tests

Resolves: DCMAW-8582
