+++ .github/workflows/on_push-develop.yml
@@ -63,16 +63,16 @@ jobs:
       - name: Run gradle unit tests
         uses: ./config/actions/gradle-check
         with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          github-actor: ${{ secrets.GITHUB_ACTOR }}
+          github-token: ${{ secrets.MODULE_FETCH_TOKEN }}
+          github-actor: ${{ secrets.MODULE_FETCH_TOKEN_USERNAME }}
           version-code: ${{ steps.version-code.outputs.version-code }}
           version-name: "${{ steps.latest-tag.outputs.current-tag }}-${{ steps.version-code.outputs.version-code }}"
 
       - name: Run gradle instrumentation tests
         uses: ./config/actions/gradle-connected-test-managed
         with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          github-actor: ${{ secrets.GITHUB_ACTOR }}
+          github-token: ${{ secrets.MODULE_FETCH_TOKEN }}
+          github-actor: ${{ secrets.MODULE_FETCH_TOKEN_USERNAME }}
           version-code: ${{ steps.version-code.outputs.version-code }}
           version-name: "${{ steps.latest-tag.outputs.current-tag }}-${{ steps.version-code.outputs.version-code }}"
